+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Hyperparameters columns in cp list are customizable (#453)
 -   Hyperparameters modal now draggable (#467)
 -   Sticky columns on compute plans page (#468)
+-   Hyperparameters modal now resizable (#475)
 
 ### Fixed